Refactor backend architecture and fix various issues
- Split monolithic `backend/main.py` into modular routers (`issues`, `detection`, `grievances`, `utility`).
- Move shared logic to `backend/utils.py`, `backend/tasks.py`, and `backend/dependencies.py`.
- Fix Python import scoping bug in `SHARED_HTTP_CLIENT` usage for cached detection functions.
- Implement atomic upvotes for issues to prevent race conditions.
- Fix broken tests (`test_local_ml_service`, `test_vandalism`, `test_smart_scan`, `test_bot_integration`) to align with refactored structure.
- Improve `stop_bot_thread` logic to ensure clean state reset.
- Suppress deprecated `google.generativeai` FutureWarning.
- Ensure all 99 tests pass.
