+++ backend/routers/issues.py
@@ -1,0 +1,533 @@
+from fastapi import APIRouter, Depends, HTTPException, UploadFile, File, Form, Query, Request, BackgroundTasks, status
+from fastapi.responses import JSONResponse
+from fastapi.concurrency import run_in_threadpool
+from sqlalchemy.orm import Session, defer
+from typing import List, Union, Dict, Any
+import uuid
+import os
+import logging
+from datetime import datetime, timezone
+
+from backend.database import get_db
+from backend.models import Issue, PushSubscription
+from backend.schemas import (
+    IssueCreateWithDeduplicationResponse, IssueCategory, NearbyIssueResponse,
+    DeduplicationCheckResponse, IssueSummaryResponse, VoteResponse,
+    IssueStatusUpdateRequest, IssueStatusUpdateResponse, PushSubscriptionRequest,
+    PushSubscriptionResponse
+)
+from backend.utils import (
+    check_upload_limits, validate_uploaded_file, save_file_blocking, save_issue_db,
+    UPLOAD_LIMIT_PER_USER, UPLOAD_LIMIT_PER_IP
+)
+from backend.tasks import (
+    process_action_plan_background, create_grievance_from_issue_background,
+    send_status_notification
+)
+from backend.spatial_utils import get_bounding_box, find_nearby_issues
+from backend.cache import recent_issues_cache
+from backend.hf_api_service import verify_resolution_vqa
+from backend.dependencies import get_http_client
+
+logger = logging.getLogger(__name__)
+
+router = APIRouter()
+
+@router.post("/api/issues", response_model=IssueCreateWithDeduplicationResponse, status_code=201)
+async def create_issue(
+    request: Request,
+    background_tasks: BackgroundTasks,
+    description: str = Form(..., min_length=10, max_length=1000),
+    category: str = Form(..., pattern=f"^({'|'.join([cat.value for cat in IssueCategory])})$"),
+    language: str = Form('en'),
+    user_email: str = Form(None),
+    latitude: float = Form(None, ge=-90, le=90),
+    longitude: float = Form(None, ge=-180, le=180),
+    location: str = Form(None, max_length=200),
+    image: UploadFile = File(None),
+    db: Session = Depends(get_db)
+):
+    image_path = None
+
+    # Check upload limits if image is being uploaded
+    if image:
+        # Get client IP
+        client_ip = request.client.host if request.client else "unknown"
+        # Use user_email if provided, otherwise IP
+        identifier = user_email if user_email else client_ip
+        limit = UPLOAD_LIMIT_PER_USER if user_email else UPLOAD_LIMIT_PER_IP
+        check_upload_limits(identifier, limit)
+
+    try:
+        # Validate uploaded image if provided
+        if image:
+            await validate_uploaded_file(image)
+
+        # Save image if provided
+        if image:
+            upload_dir = "data/uploads"
+            os.makedirs(upload_dir, exist_ok=True)
+            filename = f"{uuid.uuid4()}_{image.filename}"
+            image_path = os.path.join(upload_dir, filename)
+            await run_in_threadpool(save_file_blocking, image.file, image_path)
+    except HTTPException:
+        # Re-raise HTTP exceptions (from validation)
+        raise
+    except OSError as e:
+        logger.error(f"File I/O error while saving image: {e}", exc_info=True)
+        raise HTTPException(status_code=500, detail="Failed to save uploaded file")
+    except Exception as e:
+        logger.error(f"Unexpected error during file processing: {e}", exc_info=True)
+        raise HTTPException(status_code=500, detail="Internal server error")
+
+    # Spatial deduplication check
+    deduplication_info = None
+    linked_issue_id = None
+
+    if latitude is not None and longitude is not None:
+        try:
+            # Find existing open issues within 50 meters
+            # Optimization: Use bounding box to filter candidates in SQL
+            min_lat, max_lat, min_lon, max_lon = get_bounding_box(latitude, longitude, 50.0)
+
+            open_issues = await run_in_threadpool(
+                lambda: db.query(Issue).filter(
+                    Issue.status == "open",
+                    Issue.latitude >= min_lat,
+                    Issue.latitude <= max_lat,
+                    Issue.longitude >= min_lon,
+                    Issue.longitude <= max_lon
+                ).all()
+            )
+
+            nearby_issues_with_distance = find_nearby_issues(
+                open_issues, latitude, longitude, radius_meters=50.0
+            )
+
+            if nearby_issues_with_distance:
+                # Found nearby issues - prepare deduplication response
+                nearby_responses = [
+                    NearbyIssueResponse(
+                        id=issue.id,
+                        description=issue.description[:100] + "..." if len(issue.description) > 100 else issue.description,
+                        category=issue.category,
+                        latitude=issue.latitude,
+                        longitude=issue.longitude,
+                        distance_meters=distance,
+                        upvotes=issue.upvotes or 0,
+                        created_at=issue.created_at,
+                        status=issue.status
+                    )
+                    for issue, distance in nearby_issues_with_distance[:3]  # Limit to top 3 closest
+                ]
+
+                deduplication_info = DeduplicationCheckResponse(
+                    has_nearby_issues=True,
+                    nearby_issues=nearby_responses,
+                    recommended_action="upvote_existing"
+                )
+
+                # Automatically upvote the closest issue and link this report to it
+                closest_issue, _ = nearby_issues_with_distance[0]
+                closest_issue.upvotes = (closest_issue.upvotes or 0) + 1
+                linked_issue_id = closest_issue.id
+
+                # Update the database with the upvote
+                await run_in_threadpool(db.commit)
+
+                logger.info(f"Spatial deduplication: Linked new report to existing issue {closest_issue.id}, upvoted to {closest_issue.upvotes}")
+
+        except Exception as e:
+            logger.error(f"Error during spatial deduplication check: {e}", exc_info=True)
+            # Continue with issue creation if deduplication fails
+
+    try:
+        # Save to DB only if no nearby issues found or deduplication failed
+        if deduplication_info is None or not deduplication_info.has_nearby_issues:
+            new_issue = Issue(
+                reference_id=str(uuid.uuid4()),
+                description=description,
+                category=category,
+                image_path=image_path,
+                source="web",
+                user_email=user_email,
+                latitude=latitude,
+                longitude=longitude,
+                location=location,
+                action_plan=None
+            )
+
+            # Offload blocking DB operations to threadpool
+            await run_in_threadpool(save_issue_db, db, new_issue)
+        else:
+            # Don't create new issue, just return deduplication info
+            new_issue = None
+    except Exception as e:
+        # Clean up uploaded file if DB save failed
+        if image_path and os.path.exists(image_path):
+            try:
+                os.remove(image_path)
+            except OSError:
+                pass  # Ignore cleanup errors
+
+        logger.error(f"Database error while creating issue: {e}", exc_info=True)
+        raise HTTPException(status_code=500, detail="Failed to save issue to database")
+
+    # Add background task for AI generation only if new issue was created
+    if new_issue:
+        background_tasks.add_task(process_action_plan_background, new_issue.id, description, category, language, image_path)
+
+        # Create grievance for escalation management
+        background_tasks.add_task(create_grievance_from_issue_background, new_issue.id)
+
+        # Optimistic Cache Update with thread-safe operations
+        try:
+            current_cache = recent_issues_cache.get("recent_issues")
+            if current_cache:
+                # Create a dict representation of the new issue
+                # Using IssueSummaryResponse to match the list view optimization
+                new_issue_dict = IssueSummaryResponse(
+                    id=new_issue.id,
+                    category=new_issue.category,
+                    description=new_issue.description[:100] + "..." if len(new_issue.description) > 100 else new_issue.description,
+                    created_at=new_issue.created_at,
+                    image_path=new_issue.image_path,
+                    status=new_issue.status,
+                    upvotes=new_issue.upvotes if new_issue.upvotes is not None else 0,
+                    location=new_issue.location,
+                    latitude=new_issue.latitude,
+                    longitude=new_issue.longitude
+                    # action_plan excluded
+                ).model_dump(mode='json')
+
+                # Prepend new issue to the list (atomic operation)
+                current_cache.insert(0, new_issue_dict)
+
+                # Keep only last 10 entries
+                if len(current_cache) > 10:
+                    current_cache.pop()
+
+                # Atomic cache update
+                recent_issues_cache.set(current_cache, "recent_issues")
+        except Exception as e:
+            logger.error(f"Error updating cache optimistically: {e}")
+            # Failure to update cache is not critical, don't fail the request
+
+    # Prepare deduplication info if not already set
+    if deduplication_info is None:
+        deduplication_info = DeduplicationCheckResponse(
+            has_nearby_issues=False,
+            nearby_issues=[],
+            recommended_action="create_new"
+        )
+
+    # Return response with deduplication information
+    if new_issue:
+        return IssueCreateWithDeduplicationResponse(
+            id=new_issue.id,
+            message="Issue reported successfully. Action plan will be generated shortly.",
+            action_plan=None,
+            deduplication_info=deduplication_info,
+            linked_issue_id=linked_issue_id
+        )
+    else:
+        return IssueCreateWithDeduplicationResponse(
+            id=None,
+            message="Similar issue found nearby. Your report has been linked to the existing issue to increase its priority.",
+            action_plan=None,
+            deduplication_info=deduplication_info,
+            linked_issue_id=linked_issue_id
+        )
+
+@router.post("/api/issues/{issue_id}/vote", response_model=VoteResponse)
+def upvote_issue(issue_id: int, db: Session = Depends(get_db)):
+    issue = db.query(Issue).filter(Issue.id == issue_id).first()
+    if not issue:
+        raise HTTPException(status_code=404, detail="Issue not found")
+
+    # Increment upvotes atomically
+    if issue.upvotes is None:
+        issue.upvotes = 0
+
+    # Use SQLAlchemy expression for atomic update
+    issue.upvotes = Issue.upvotes + 1
+
+    db.commit()
+    db.refresh(issue)
+
+    return VoteResponse(
+        id=issue.id,
+        upvotes=issue.upvotes,
+        message="Issue upvoted successfully"
+    )
+
+@router.get("/api/issues/nearby", response_model=List[NearbyIssueResponse])
+def get_nearby_issues(
+    latitude: float = Query(..., ge=-90, le=90, description="Latitude of the location"),
+    longitude: float = Query(..., ge=-180, le=180, description="Longitude of the location"),
+    radius: float = Query(50.0, ge=10, le=500, description="Search radius in meters"),
+    limit: int = Query(10, ge=1, le=50, description="Maximum number of results"),
+    db: Session = Depends(get_db)
+):
+    """
+    Get issues near a specific location for deduplication purposes.
+    Returns issues within the specified radius, sorted by distance.
+    """
+    try:
+        # Query open issues with coordinates
+        # Optimization: Use bounding box to filter candidates in SQL
+        min_lat, max_lat, min_lon, max_lon = get_bounding_box(latitude, longitude, radius)
+
+        open_issues = db.query(Issue).filter(
+            Issue.status == "open",
+            Issue.latitude >= min_lat,
+            Issue.latitude <= max_lat,
+            Issue.longitude >= min_lon,
+            Issue.longitude <= max_lon
+        ).all()
+
+        nearby_issues_with_distance = find_nearby_issues(
+            open_issues, latitude, longitude, radius_meters=radius
+        )
+
+        # Convert to response format and limit results
+        nearby_responses = [
+            NearbyIssueResponse(
+                id=issue.id,
+                description=issue.description[:100] + "..." if len(issue.description) > 100 else issue.description,
+                category=issue.category,
+                latitude=issue.latitude,
+                longitude=issue.longitude,
+                distance_meters=distance,
+                upvotes=issue.upvotes or 0,
+                created_at=issue.created_at,
+                status=issue.status
+            )
+            for issue, distance in nearby_issues_with_distance[:limit]
+        ]
+
+        return nearby_responses
+
+    except Exception as e:
+        logger.error(f"Error getting nearby issues: {e}", exc_info=True)
+        raise HTTPException(status_code=500, detail="Failed to retrieve nearby issues")
+
+@router.post("/api/issues/{issue_id}/verify", response_model=Union[VoteResponse, Dict[str, Any]])
+async def verify_issue_endpoint(
+    issue_id: int,
+    request: Request,
+    image: UploadFile = File(None),
+    db: Session = Depends(get_db)
+):
+    issue = await run_in_threadpool(lambda: db.query(Issue).filter(Issue.id == issue_id).first())
+    if not issue:
+        raise HTTPException(status_code=404, detail="Issue not found")
+
+    if image:
+        # AI Verification Logic
+        # Validate uploaded file
+        await validate_uploaded_file(image)
+
+        try:
+            image_bytes = await image.read()
+        except Exception as e:
+            logger.error(f"Invalid image file: {e}", exc_info=True)
+            raise HTTPException(status_code=400, detail="Invalid image file")
+
+        # Construct question
+        category = issue.category.lower() if issue.category else "issue"
+        question = f"Is there a {category} in this image?"
+
+        # Custom questions for common categories
+        if "pothole" in category:
+            question = "Is there a pothole on the road?"
+        elif "garbage" in category or "waste" in category:
+            question = "Is there garbage or trash on the ground?"
+        elif "light" in category:
+            question = "Is the streetlight broken?"
+        elif "water" in category or "flood" in category:
+            question = "Is the street flooded?"
+        elif "tree" in category:
+            question = "Is there a fallen tree?"
+
+        try:
+            # Use shared client dependency is tricky here because logic is mixed
+            # request.app.state.http_client is available
+            client = request.app.state.http_client
+            result = await verify_resolution_vqa(image_bytes, question, client)
+
+            answer = result.get('answer', 'unknown')
+            confidence = result.get('confidence', 0)
+
+            # If the answer is "no" (meaning the issue is NOT present), we consider it resolved.
+            is_resolved = False
+            if answer.lower() in ["no", "none", "nothing"] and confidence > 0.5:
+                is_resolved = True
+                # Update status if not already resolved
+                if issue.status != "resolved":
+                    issue.status = "verified" # Mark as verified (resolved usually implies closed)
+                    issue.verified_at = datetime.now(timezone.utc)
+                    await run_in_threadpool(db.commit)
+
+            return {
+                "is_resolved": is_resolved,
+                "ai_answer": answer,
+                "confidence": confidence,
+                "question_asked": question
+            }
+        except Exception as e:
+            logger.error(f"Resolution verification error: {e}", exc_info=True)
+            raise HTTPException(status_code=500, detail="Verification service temporarily unavailable")
+    else:
+        # Manual Verification Logic (Vote)
+        # Increment upvotes (verification counts as strong support)
+        if issue.upvotes is None:
+            issue.upvotes = 0
+
+        # Atomic increment
+        issue.upvotes = Issue.upvotes + 2
+
+        # If issue has enough verifications, consider upgrading status
+        # Note: We can't check issue.upvotes value immediately if it's an expression
+        # We need to refresh first, but we want to do logic based on new value.
+        # This is tricky with atomic updates.
+        # For now, let's commit and refresh to get value, then update status if needed.
+
+        await run_in_threadpool(db.commit)
+        await run_in_threadpool(db.refresh, issue)
+
+        if issue.upvotes >= 5 and issue.status == "open":
+            issue.status = "verified"
+            logger.info(f"Issue {issue_id} automatically verified due to {issue.upvotes} upvotes")
+            await run_in_threadpool(db.commit) # Commit status change
+
+        return VoteResponse(
+            id=issue.id,
+            upvotes=issue.upvotes,
+            message="Issue verified successfully"
+        )
+
+@router.put("/api/issues/status", response_model=IssueStatusUpdateResponse)
+def update_issue_status(
+    request: IssueStatusUpdateRequest,
+    background_tasks: BackgroundTasks,
+    db: Session = Depends(get_db)
+):
+    """Update issue status via secure reference ID (for government portals)"""
+    issue = db.query(Issue).filter(Issue.reference_id == request.reference_id).first()
+    if not issue:
+        raise HTTPException(status_code=404, detail="Issue not found")
+
+    # Validate status transition (simple state machine)
+    valid_transitions = {
+        "open": ["verified"],
+        "verified": ["assigned", "open"],
+        "assigned": ["in_progress", "verified"],
+        "in_progress": ["resolved", "assigned"],
+        "resolved": []  # Terminal state
+    }
+
+    if request.status.value not in valid_transitions.get(issue.status, []):
+        raise HTTPException(
+            status_code=400,
+            detail=f"Invalid status transition from {issue.status} to {request.status.value}"
+        )
+
+    # Update issue
+    old_status = issue.status
+    issue.status = request.status.value
+    if request.assigned_to:
+        issue.assigned_to = request.assigned_to
+
+    # Set timestamps
+    now = datetime.now(timezone.utc)
+    if request.status.value == "verified":
+        issue.verified_at = now
+    elif request.status.value == "assigned":
+        issue.assigned_at = now
+    elif request.status.value == "resolved":
+        issue.resolved_at = now
+
+    db.commit()
+    db.refresh(issue)
+
+    # Send notification to citizen
+    background_tasks.add_task(send_status_notification, issue.id, old_status, request.status.value, request.notes)
+
+    return IssueStatusUpdateResponse(
+        id=issue.id,
+        reference_id=issue.reference_id,
+        status=request.status,
+        message=f"Issue status updated to {request.status.value}"
+    )
+
+@router.post("/api/push-subscription", response_model=PushSubscriptionResponse)
+def subscribe_push_notifications(
+    request: PushSubscriptionRequest,
+    db: Session = Depends(get_db)
+):
+    """Subscribe to push notifications for issue updates"""
+    # Check if subscription already exists
+    existing = db.query(PushSubscription).filter(
+        PushSubscription.endpoint == request.endpoint
+    ).first()
+
+    if existing:
+        # Update existing subscription
+        existing.user_email = request.user_email
+        existing.p256dh = request.p256dh
+        existing.auth = request.auth
+        existing.issue_id = request.issue_id
+        db.commit()
+        return PushSubscriptionResponse(
+            id=existing.id,
+            message="Push subscription updated"
+        )
+
+    # Create new subscription
+    subscription = PushSubscription(
+        user_email=request.user_email,
+        endpoint=request.endpoint,
+        p256dh=request.p256dh,
+        auth=request.auth,
+        issue_id=request.issue_id
+    )
+
+    db.add(subscription)
+    db.commit()
+    db.refresh(subscription)
+
+    return PushSubscriptionResponse(
+        id=subscription.id,
+        message="Push subscription created"
+    )
+
+@router.get("/api/issues/recent", response_model=List[IssueSummaryResponse])
+def get_recent_issues(db: Session = Depends(get_db)):
+    cached_data = recent_issues_cache.get("recent_issues")
+    if cached_data:
+        return JSONResponse(content=cached_data)
+
+    # Fetch last 10 issues, deferring action_plan for performance
+    issues = db.query(Issue).options(defer(Issue.action_plan)).order_by(Issue.created_at.desc()).limit(10).all()
+
+    # Convert to Pydantic models for validation and serialization
+    data = []
+    for i in issues:
+        data.append(IssueSummaryResponse(
+            id=i.id,
+            category=i.category,
+            description=i.description[:100] + "..." if len(i.description) > 100 else i.description,
+            created_at=i.created_at,
+            image_path=i.image_path,
+            status=i.status,
+            upvotes=i.upvotes if i.upvotes is not None else 0,
+            location=i.location,
+            latitude=i.latitude,
+            longitude=i.longitude
+            # action_plan is deferred and excluded
+        ).model_dump(mode='json'))
+
+    # Thread-safe cache update
+    recent_issues_cache.set(data, "recent_issues")
+    return data
