+++ backend/tests/test_detection_bytes.py
@@ -90,9 +90,9 @@ async def test_detect_vandalism_with_bytes(client):
     img_bytes = img_byte_arr.getvalue()
 
     # Send request
-    with patch('backend.main.validate_uploaded_file'), \
-         patch('backend.main.validate_image_for_processing'), \
-         patch('backend.main.detect_vandalism_unified', AsyncMock(return_value=[{"label": "graffiti", "score": 0.95}])):
+    with patch('backend.utils.validate_uploaded_file'), \
+         patch('backend.pothole_detection.validate_image_for_processing'), \
+         patch('backend.routers.detection.detect_vandalism_unified', AsyncMock(return_value=[{"label": "graffiti", "score": 0.95}])):
         response = client.post(
             "/api/detect-vandalism",
             files={"image": ("test.jpg", img_bytes, "image/jpeg")}
@@ -128,9 +128,9 @@ async def test_detect_infrastructure_with_bytes(client):
     img.save(img_byte_arr, format='JPEG')
     img_bytes = img_byte_arr.getvalue()
 
-    with patch('backend.main.validate_uploaded_file'), \
-         patch('backend.main.validate_image_for_processing'), \
-         patch('backend.main.detect_infrastructure_unified', AsyncMock(return_value=[{"label": "fallen tree", "score": 0.8}])):
+    with patch('backend.utils.validate_uploaded_file'), \
+         patch('backend.pothole_detection.validate_image_for_processing'), \
+         patch('backend.routers.detection.detect_infrastructure_unified', AsyncMock(return_value=[{"label": "fallen tree", "score": 0.8}])):
         response = client.post(
             "/api/detect-infrastructure",
             files={"image": ("test.jpg", img_bytes, "image/jpeg")}
