+++ tests/test_captioning.py
@@ -1,12 +1,12 @@
 from fastapi.testclient import TestClient
 from unittest.mock import patch, AsyncMock
-from main import app
+from backend.main import app
 import pytest
 
 @pytest.mark.asyncio
 async def test_generate_description_endpoint():
-    # Mock the generate_image_caption function in 'main' module
-    with patch("main.generate_image_caption", new_callable=AsyncMock) as mock_caption:
+    # Mock the generate_image_caption function in 'backend.routers.detection' module
+    with patch("backend.routers.detection.generate_image_caption", new_callable=AsyncMock) as mock_caption:
         mock_caption.return_value = "A photo of a pothole on the road"
 
         with TestClient(app) as client:
