+++ tests/test_issue_creation.py
@@ -24,16 +24,16 @@ def test_create_issue():
     # Create a dummy image file (valid JPEG header)
     with tempfile.NamedTemporaryFile(suffix=".jpg", delete=False) as tmp:
         # Minimal JPEG header
-        tmp.write(b'\xff\xd8\xff\xe0\x00\x10\x4a\x46\x49\x46\x00\x01\x01\x01\x00\x48\x00\x48\x00\x00\xff\xdb\x00\x43\x00\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xc0\x00\x0b\x08\x00\x01\x00\x01\x01\x01\x11\x00\xff\xc4\x00\x1f\x00\x00\x01\x05\x01\x01\x01\x01\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x0a\x0b\xff\xda\x00\x08\x01\x01\x00\x00\x3f\x00\x7f\xff\xd9')
+        tmp.write(b'\xff\xd8\xff\xe0\x00\x10\x4a\x46\x49\x46\x00\x01\x01\x01\x00\x48\x00\x48\x00\x00\xff\xdb\x00\x43\x00\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xc0\x00\x0b\x08\x00\x01\x01\x01\x11\x00\xff\xc4\x00\x1f\x00\x00\x01\x05\x01\x01\x01\x01\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x0a\x0b\xff\xda\x00\x08\x01\x01\x00\x00\x3f\x00\x7f\xff\xd9')
         tmp_path = tmp.name
 
     try:
         from unittest.mock import patch, AsyncMock
         # Patch validation to avoid PIL/magic issues with dummy image
         # Also patch action plan generation to avoid external API calls
-        # Note: Patch where it is imported/used (backend.main), not where it is defined
-        with patch("backend.main.validate_uploaded_file", new_callable=AsyncMock) as mock_validate, \
-             patch("backend.main.generate_action_plan", new_callable=AsyncMock) as mock_plan:
+        # Note: Patch where it is imported/used (backend.routers.issues and backend.tasks)
+        with patch("backend.routers.issues.validate_uploaded_file", new_callable=AsyncMock) as mock_validate, \
+             patch("backend.tasks.generate_action_plan", new_callable=AsyncMock) as mock_plan:
 
             mock_plan.return_value = {
                 "whatsapp": "Test WhatsApp",
