+++ tests/test_vandalism.py
@@ -17,10 +17,10 @@ def test_read_main(client):
     assert "data" in json_response
     assert json_response["data"]["service"] == "VishwaGuru API"
 
-@patch("backend.main.magic.from_buffer")
-@patch("backend.main.detect_vandalism_local", new_callable=AsyncMock)
-@patch("backend.main.run_in_threadpool")
-@patch("backend.main.Image.open")
+@patch("backend.utils.magic.from_buffer")
+@patch("backend.routers.detection.detect_vandalism_unified", new_callable=AsyncMock)
+@patch("backend.utils.run_in_threadpool")
+@patch("backend.utils.Image.open")
 def test_detect_vandalism(mock_image_open, mock_run, mock_detect_vandalism, mock_magic, client):
     # Mock magic to return image/jpeg
     mock_magic.return_value = "image/jpeg"
@@ -37,8 +37,11 @@ def test_detect_vandalism(mock_image_open, mock_run, mock_detect_vandalism, mock
     mock_detect_vandalism.return_value = mock_result
 
     # Note: run_in_threadpool is still used for Image.open, so we mock it
-    # But for detection it is NOT used.
+    # But for detection it is NOT used directly in the test scope but inside utils
     async def async_mock_run_img(*args, **kwargs):
+        # Return whatever was passed if it's the function, or mock image if it's open
+        if args and args[0] == mock_image.verify:
+             return None
         return mock_image
 
     mock_run.side_effect = async_mock_run_img
